V1.4.2

+++ app/src/main/java/com/projeto/bookfast/bookfast/dominio/Pessoa.java
@@ -1,9 +1,11 @@
 package com.projeto.bookfast.bookfast.dominio;
 
 
+import android.text.Editable;
+
 public class Pessoa {
-    private long id;
-    private long cpf;
+    private int id;
+    private int cpf;
     private String nome;
     private String email;
     private String senha;
@@ -14,7 +16,7 @@ public class Pessoa {
 
     }
     //Construtor para atualizar
-    public Pessoa(long _id, long _cpf, String _nome, String _email, String _senha) {
+    public Pessoa(int _id, int _cpf, String _nome, String _email, String _senha) {
         this.id = _id;
         this.cpf = _cpf;
         this.nome = _nome;
@@ -22,7 +24,7 @@ public class Pessoa {
         this.senha = _senha;
     }
     //Construtor para inserir
-    public Pessoa(long _cpf, String _nome, String _email, String _senha) {
+    public Pessoa(int _cpf, String _nome, String _email, String _senha) {
         this.cpf = _cpf;
         this.nome = _nome;
         this.email = _email;
@@ -30,19 +32,19 @@ public class Pessoa {
     }
     //>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
-    public long getCpf() {
+    public int getCpf() {
         return cpf;
     }
 
-    public void setCpf(long cpf) {
+    public void setCpf(int cpf) {
         this.cpf = cpf;
     }
 
@@ -69,4 +71,6 @@ public class Pessoa {
     public void setSenha(String senha) {
         this.senha = senha;
     }
+
+
 }
